# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 知行创境教育创作平台 API
  version: 1.0.0
  description: |
    基于《知行创境》教育创作平台 V2.0 PRD 整理的后端接口 OpenAPI 文档。
    基础路径为 `/api/v1`，认证方式为 Bearer Token。
servers:
  - url: /api/v1
    description: Default API Server
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginData'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /auth/me:
    get:
      tags: [Auth]
      summary: 获取当前用户信息
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 当前用户信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /classes:
    get:
      tags: [Classes]
      summary: 获取我的班级列表
      operationId: getMyClasses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 班级列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ClassInfo'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /classes/{classId}:
    get:
      tags: [Classes]
      summary: 获取班级详情
      operationId: getClassDetail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: 班级详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ClassDetail'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /classes/{classId}/members:
    get:
      tags: [Classes]
      summary: 获取班级成员
      operationId: getClassMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: 班级成员列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ClassMember'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /classes/{classId}/announcements:
    get:
      tags: [Classes]
      summary: 获取班级公告列表
      operationId: getClassAnnouncements
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: 公告列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Announcement'
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Classes]
      summary: 发布班级公告
      operationId: createAnnouncement
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClassId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnnouncementRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Announcement'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /assets/upload:
    post:
      tags: [Assets]
      summary: 上传素材
      operationId: uploadAsset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                folder:
                  type: string
                  example: exhibition
                bizType:
                  type: string
                  example: exhibition_material
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AssetUploadData'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /assets:
    get:
      tags: [Assets]
      summary: 获取我的素材列表
      operationId: getAssets
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: assetType
          schema:
            $ref: '#/components/schemas/AssetType'
        - in: query
          name: sourceType
          schema:
            $ref: '#/components/schemas/AssetSourceType'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 素材分页列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedAsset'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ClassId:
      in: path
      name: classId
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
...
    ErrorResponse:
      description: 业务异常
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: 40100
                message: 未登录或 token 失效
                data: null
                requestId: 4f9b6e24e2ef4f4f9d6c
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
        requestId:
          type: string
      required: [code, message, requestId]
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 40001
        message:
          type: string
          example: 参数错误
        data:
          nullable: true
        requestId:
          type: string
      required: [code, message, requestId]
    UserRole:
      type: string
      enum: [student, teacher, parent, admin, visitor]
...
    AssetType:
      type: string
      enum: [image, video, audio, document, model]
    AssetSourceType:
      type: string
      enum: [upload, platform, museum]
    SchoolInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    ClassInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        grade:
          type: string
        academicYear:
          type: string
    SimpleUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          $ref: '#/components/schemas/UserRole'
        name:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          format: uri
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          $ref: '#/components/schemas/UserRole'
        realName:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          format: uri
        email:
          type: string
          format: email
        mobile:
          type: string
        bio:
          type: string
        school:
          $ref: '#/components/schemas/SchoolInfo'
        classInfo:
          $ref: '#/components/schemas/ClassInfo'
    ClassDetail:
      allOf:
        - $ref: '#/components/schemas/ClassInfo'
        - type: object
          properties:
            school:
              $ref: '#/components/schemas/SchoolInfo'
            headTeacher:
              $ref: '#/components/schemas/SimpleUser'
            memberCount:
              type: integer
    ClassMember:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        role:
          type: string
          enum: [student, teacher, assistant]
        realName:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          format: uri
        joinedAt:
          type: string
          format: date-time
    Announcement:
      type: object
      properties:
        id:
          type: integer
          format: int64
        classId:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        pinned:
          type: boolean
        publisher:
          $ref: '#/components/schemas/SimpleUser'
        publishedAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [account, password]
      properties:
        account:
          type: string
        password:
          type: string
          format: password
    LoginData:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/SimpleUser'
...
    CreateAnnouncementRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        pinned:
          type: boolean
          default: false
...
    AssetUploadData:
      type: object
      properties:
        assetId:
          type: integer
          format: int64
        fileName:
          type: string
        fileUrl:
          type: string
          format: uri
        mimeType:
          type: string
        fileSize:
          type: integer
          format: int64
    Asset:
      allOf:
        - $ref: '#/components/schemas/AssetUploadData'
        - type: object
          properties:
            assetType:
              $ref: '#/components/schemas/AssetType'
            sourceType:
              $ref: '#/components/schemas/AssetSourceType'
            createdAt:
              type: string
              format: date-time
...
    PagedAsset:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...